first of the project
Hello Internet, this is Gavin speaking.

This app is still NOT FINISH YET, if you unfortuately see this project in your searching list, plz ignore it as soon as possible.

This diary app is referenced to the app which shown in the anime movie "Your Name", or maybe "Kimi no na wa" will ring a bell? Anyway, this app might be finished before the end of this semester, that means June or May, make sure to keep following if you are interested with this ;)